2021/7/30 完成  seate 分布式事务模块,以及素材管理(增删查改) fastDFS

+++ heima-leadnews-common/src/main/java/com/heima/common/constans/user/UserConstants.java
@@ -8,7 +8,7 @@ import java.security.PublicKey;
  * @ Description：
  */
 public class UserConstants {
-   public static final Short FAIL_AUTH=2;
-   public static final Short PASS_AUTH=9;
+   public static final Integer FAIL_AUTH=2;
+   public static final Integer PASS_AUTH=9;
    public static final Integer AUTH_TYPE=2;
 }
